+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [Unreleased] 
+
+### Added
+
+- [TD-2591] Include the number of related concepts on concept view
 
 ### Changed